feat: Add Numba environment configuration and context managers for parallel execution
