+++ brainevent/_event_matrix_impl.py
@@ -21,10 +21,11 @@ import jax.numpy as jnp
 import numpy as np
 from jax.interpreters import ad
 
+from ._config import numba_environ
 from ._misc import cdiv
 from ._typing import Kernel
 from ._xla_custom_op import XLACustomKernel
-from ._xla_custom_op_numba import numba_environ, NumbaKernelGenerator
+from ._xla_custom_op_numba import NumbaKernelGenerator
 from ._xla_custom_op_util import general_batching_rule
 from ._xla_custom_op_warp import WarpKernelGenerator, dtype_to_warp_type
