+++ brainevent/_jitc_float_homo_impl_test.py
@@ -64,7 +64,7 @@ class TestJitcCsrMatvecHomo:
     @pytest.mark.parametrize('shape', [(20, 20), (100, 50)])
     @pytest.mark.parametrize('transpose', [True, False])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_todense(self, shape: MatrixShape,  transpose, corder):
+    def test_todense(self, shape: MatrixShape, transpose, corder):
         jitc = brainevent.JITCHomoR(
             (1.5, 0.1, 123),
             shape=shape,
@@ -109,7 +109,7 @@ class TestJitcCsrMatvecHomo:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('transpose', [True, False])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_random_connectivity(self, shape: MatrixShape,  weight, prob, transpose, corder):
+    def test_random_connectivity(self, shape: MatrixShape, weight, prob, transpose, corder):
         seed = 1234
         vector = jnp.asarray(np.random.rand(shape[0] if transpose else shape[1]))
         r1 = float_jitc_homo_matvec(
@@ -179,7 +179,7 @@ class TestJitcCsrMatmatHomo:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('transpose', [True, False])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_random_connectivity(self, shape: MatrixShape,  batch_size, weight, prob, transpose, corder):
+    def test_random_connectivity(self, shape: MatrixShape, batch_size, weight, prob, transpose, corder):
         seed = 1234
 
         print(
@@ -297,7 +297,7 @@ class Test_JITCHomoR:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('weight', [1.5, 2.1 * u.mV])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
-    def test_jitmat(self, prob, weight, shape: MatrixShape,  k):
+    def test_jitmat(self, prob, weight, shape: MatrixShape, k):
         jitc = brainevent.JITCHomoR((weight, prob, 123), shape=shape)
         matrix = jnp.asarray(np.random.rand(shape[1], k))
         out1 = jitc @ matrix
@@ -592,7 +592,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_vector(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_vector(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(batch_size, shape[1])
 
         def f(vector):
@@ -610,7 +610,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_vector_axis1(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_vector_axis1(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(shape[1], batch_size)
 
         def f(vector):
@@ -628,7 +628,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_weight(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_weight(self, batch_size, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         vector = brainstate.random.rand(shape[1], )
 
@@ -647,7 +647,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_vector(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_vector(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(batch_size, shape[0])
 
         def f(vector):
@@ -665,7 +665,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_vector_axis1(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_vector_axis1(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(shape[0], batch_size)
 
         def f(vector):
@@ -683,7 +683,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_weight(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_weight(self, batch_size, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         vector = brainstate.random.rand(shape[0], )
 
@@ -703,7 +703,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(batch_size, shape[1], k)
 
         def f(mat):
@@ -722,7 +722,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(shape[1], batch_size, k)
 
         def f(mat):
@@ -741,7 +741,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(shape[1], k, batch_size, )
 
         def f(mat):
@@ -760,7 +760,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_weight(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_weight(self, batch_size, k, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         matrix = brainstate.random.rand(shape[1], k)
 
@@ -780,7 +780,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(batch_size, k, shape[0])
 
         def f(mat):
@@ -799,7 +799,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(k, batch_size, shape[0])
 
         def f(mat):
@@ -818,7 +818,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(k, shape[0], batch_size)
 
         def f(mat):
@@ -837,7 +837,7 @@ class Test_JITCHomoR_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_weight(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_weight(self, batch_size, k, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         mat = brainstate.random.rand(k, shape[0], )
 
@@ -879,7 +879,7 @@ class Test_JITCHomoR_Transpose:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('weight', [1.5, 2.1 * u.mV])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
-    def test_jitmat(self, prob, weight, shape: MatrixShape,  k):
+    def test_jitmat(self, prob, weight, shape: MatrixShape, k):
         jitc = brainevent.JITCHomoR((weight, prob, 123), shape=shape).T
         matrix = jnp.asarray(np.random.rand(shape[0], k))
         out1 = jitc @ matrix
@@ -1137,7 +1137,7 @@ class Test_JITCHomoC:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('weight', [1.5, 2.1 * u.mV])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
-    def test_jitmat(self, prob, weight, shape: MatrixShape,  k):
+    def test_jitmat(self, prob, weight, shape: MatrixShape, k):
         jitc = brainevent.JITCHomoC((weight, prob, 123), shape=shape)
         matrix = jnp.asarray(np.random.rand(shape[1], k))
         out1 = jitc @ matrix
@@ -1432,7 +1432,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_vector(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_vector(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(batch_size, shape[1])
 
         def f(vector):
@@ -1450,7 +1450,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_vector_axis1(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_vector_axis1(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(shape[1], batch_size)
 
         def f(vector):
@@ -1468,7 +1468,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matvec_batching_weight(self, batch_size, shape: MatrixShape,  corder):
+    def test_matvec_batching_weight(self, batch_size, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         vector = brainstate.random.rand(shape[1], )
 
@@ -1487,7 +1487,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_vector(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_vector(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(batch_size, shape[0])
 
         def f(vector):
@@ -1505,7 +1505,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_vector_axis1(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_vector_axis1(self, batch_size, shape: MatrixShape, corder):
         vectors = brainstate.random.rand(shape[0], batch_size)
 
         def f(vector):
@@ -1523,7 +1523,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('batch_size', [10])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_vecmat_batching_weight(self, batch_size, shape: MatrixShape,  corder):
+    def test_vecmat_batching_weight(self, batch_size, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         vector = brainstate.random.rand(shape[0], )
 
@@ -1543,7 +1543,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(batch_size, shape[1], k)
 
         def f(mat):
@@ -1562,7 +1562,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(shape[1], batch_size, k)
 
         def f(mat):
@@ -1581,7 +1581,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape, corder):
         matrices = brainstate.random.rand(shape[1], k, batch_size, )
 
         def f(mat):
@@ -1600,7 +1600,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_jitmat_batching_weight(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_jitmat_batching_weight(self, batch_size, k, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         matrix = brainstate.random.rand(shape[1], k)
 
@@ -1620,7 +1620,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(batch_size, k, shape[0])
 
         def f(mat):
@@ -1639,7 +1639,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix_axis1(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(k, batch_size, shape[0])
 
         def f(mat):
@@ -1658,7 +1658,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_matrix_axis2(self, batch_size, k, shape: MatrixShape, corder):
         matrix = brainstate.random.rand(k, shape[0], batch_size)
 
         def f(mat):
@@ -1677,7 +1677,7 @@ class Test_JITCHomoC_Batching:
     @pytest.mark.parametrize('k', [5])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
     @pytest.mark.parametrize('corder', [True, False])
-    def test_matjit_batching_weight(self, batch_size, k, shape: MatrixShape,  corder):
+    def test_matjit_batching_weight(self, batch_size, k, shape: MatrixShape, corder):
         weights = brainstate.random.rand(batch_size)
         mat = brainstate.random.rand(k, shape[0], )
 
@@ -1719,7 +1719,7 @@ class Test_JITCHomoC_Transpose:
     @pytest.mark.parametrize('prob', [0.1, 0.2])
     @pytest.mark.parametrize('weight', [1.5, 2.1 * u.mV])
     @pytest.mark.parametrize('shape', [(20, 30), (100, 50)])
-    def test_jitmat(self, prob, weight, shape: MatrixShape,  k):
+    def test_jitmat(self, prob, weight, shape: MatrixShape, k):
         jitc = brainevent.JITCHomoC((weight, prob, 123), shape=shape).T
         matrix = jnp.asarray(np.random.rand(shape[0], k))
         out1 = jitc @ matrix
