+++ brainevent/_jitc_float_uniform_impl.py
@@ -23,12 +23,12 @@ from jax import numpy as jnp
 from jax.interpreters import ad
 
 from ._compatible_import import pallas as pl
-from ._pallas_random import LFSR88RNG
+from ._config import Config, numba_environ
 from ._jitc_util import _initialize_seed, _initialize_conn_length
-from ._misc import Config
+from ._pallas_random import LFSR88RNG
 from ._typing import Kernel, Data, MatrixShape
 from ._xla_custom_op import XLACustomKernel
-from ._xla_custom_op_numba import NumbaKernelGenerator, numba_environ
+from ._xla_custom_op_numba import NumbaKernelGenerator
 from ._xla_custom_op_pallas import PallasKernelGenerator
 from ._xla_custom_op_util import general_batching_rule
 from ._xla_custom_op_warp import dtype_to_warp_type, WarpKernelGenerator
