+++ brainevent/_xla_custom_op_numba.py
@@ -19,9 +19,7 @@ import ctypes
 import dataclasses
 import functools
 import importlib.util
-import threading
-from contextlib import contextmanager
-from typing import Callable, Dict, Union
+from typing import Callable, Dict
 
 from jax.interpreters import mlir
 
@@ -29,72 +27,9 @@ from ._compatible_import import register_custom_call, Primitive, custom_call
 
 __all__ = [
     'NumbaKernelGenerator',
-    'set_numba_environ',
-    'numba_environ_context',
 ]
 
 numba_installed = importlib.util.find_spec('numba') is not None
-
-
-class NumbaEnvironment(threading.local):
-    def __init__(self, *args, **kwargs):
-        # default environment settings
-        super().__init__(*args, **kwargs)
-        self.parallel: bool = False
-        self.setting: dict = dict(nogil=True, fastmath=True)
-
-
-numba_environ = NumbaEnvironment()
-
-
-@contextmanager
-def numba_environ_context(
-    parallel_if_possible: Union[int, bool] = None,
-    **kwargs
-):
-    """
-    Enable Numba parallel execution if possible.
-    """
-    old_parallel = numba_environ.parallel
-    old_setting = numba_environ.setting.copy()
-
-    try:
-        numba_environ.setting.update(kwargs)
-        if parallel_if_possible is not None:
-            if isinstance(parallel_if_possible, bool):
-                numba_environ.parallel = parallel_if_possible
-            elif isinstance(parallel_if_possible, int):
-                numba_environ.parallel = True
-                assert parallel_if_possible > 0, 'The number of threads must be a positive integer.'
-                import numba  # pylint: disable=import-outside-toplevel
-                numba.set_num_threads(parallel_if_possible)
-            else:
-                raise ValueError('The argument `parallel_if_possible` must be a boolean or an integer.')
-        yield numba_environ.setting.copy()
-    finally:
-        numba_environ.parallel = old_parallel
-        numba_environ.setting = old_setting
-
-
-@contextmanager
-def set_numba_environ(
-    parallel_if_possible: Union[int, bool] = None,
-    **kwargs
-):
-    """
-    Enable Numba parallel execution if possible.
-    """
-    numba_environ.setting.update(kwargs)
-    if parallel_if_possible is not None:
-        if isinstance(parallel_if_possible, bool):
-            numba_environ.parallel = parallel_if_possible
-        elif isinstance(parallel_if_possible, int):
-            numba_environ.parallel = True
-            assert parallel_if_possible > 0, 'The number of threads must be a positive integer.'
-            import numba  # pylint: disable=import-outside-toplevel
-            numba.set_num_threads(parallel_if_possible)
-        else:
-            raise ValueError('The argument `parallel_if_possible` must be a boolean or an integer.')
 
 
 @dataclasses.dataclass(frozen=True)
